heavy refactor of ai-parrot
