+++ examples/next/google.py
@@ -212,27 +212,31 @@ async def main():
     #     print("- Math data:", math_response.output)
     #     print("- Parallel tools used:", len(math_response.tool_calls))
 
-    # print(' --- Video Features --- ')
-    # async with GoogleGenAIClient() as client:
-    #     # video generation:
-    #     prompt_data = VideoGenerationPrompt(
-    #         prompt="a Bison running for cover in a dense forest, a close-up shot of the bison's face, with the camera following its movement, is cold and we can see the bison's breath in the air, the camera is moving quickly to keep up with the bison, the bison is running through a dense forest with snow on the ground, the camera is following the bison's movement, the bison is running for cover in a dense forest, a close-up shot of the bison's face, with the camera following its movement, is cold and we can see the bison's breath in the air, the camera is moving quickly to keep up with the bison",
-    #         number_of_videos=1,
-    #         model=GoogleModel.VEO_3_0,
-    #         aspect_ratio="16:9",
-    #         duration=10,  # Duration in seconds
-    #     )
-    #     output_directory = BASE_DIR.joinpath('static', 'generated_videos')
-    #     output_directory.mkdir(parents=True, exist_ok=True)
+    print(' --- Video Features --- ')
+    async with GoogleGenAIClient() as client:
+        # video generation:
+        prompt_data = VideoGenerationPrompt(
+            # prompt="a Bison running for cover in a dense forest, a close-up shot of the bison's face, with the camera following its movement, is cold and we can see the bison's breath in the air, the camera is moving quickly to keep up with the bison, the bison is running through a dense forest with snow on the ground, the camera is following the bison's movement, the bison is running for cover in a dense forest, a close-up shot of the bison's face, with the camera following its movement, is cold and we can see the bison's breath in the air, the camera is moving quickly to keep up with the bison",
+            #prompt="Un peluche de gato con alas volando en un cielo azul con nubes esponjosas, el peluche tiene un diseño colorido y amigable, las alas son grandes y suaves, el cielo es brillante y soleado, el peluche parece feliz mientras vuela entre las nubes",
+            #styles=["fantasy", "whimsical"],
+            prompt="Una amable señora en un paisaje de la Toscana Italiana tejiendo crochet bajo un árbol de olivo, el sol brilla suavemente, creando un ambiente cálido y acogedor, la señora sonríe mientras trabaja en su proyecto de crochet, rodeada de campos verdes y colinas suaves",
+            styles=["photorealistic", "warm"],
+            number_of_videos=1,
+            model=GoogleModel.VEO_3_0,
+            aspect_ratio="16:9",
+            duration=10,  # Duration in seconds
+        )
+        output_directory = BASE_DIR.joinpath('static', 'generated_videos')
+        output_directory.mkdir(parents=True, exist_ok=True)
 
-    #     response = await client.generate_videos(
-    #         prompt=prompt_data,
-    #         output_directory=output_directory,
-    #         mime_format="video/mp4",
-    #     )
-    #     print("Generated videos:")
-    #     for video in response.videos:
-    #         print(f"✅ Video saved to: {video.video}")
+        response = await client.generate_videos(
+            prompt=prompt_data,
+            output_directory=output_directory,
+            mime_format="video/mp4",
+        )
+        print("Generated videos:")
+        for video in response.videos:
+            print(f"✅ Video saved to: {video.video}")
     # print(' --- Image Features --- ')
     # async with GoogleGenAIClient() as client:
     #     # Image Generation:
@@ -302,20 +306,20 @@ async def main():
 #             print("❌ Failed to generate multi-voice speech.")
 
         # Image Clasification:
-    async with GoogleGenAIClient() as client:
-        class ImageCategory(str, Enum):
-            """Enumeration for retail image categories."""
-            INK_WALL = "Ink Wall"
-            SHELVES_WITH_PRODUCTS = "Shelves with Products"
-            PRODUCTS_ON_FLOOR = "Products on Floor"
-            MERCHANDISING_ENDCAP = "Merchandising Endcap"
-            OTHER = "Other"
+    # async with GoogleGenAIClient() as client:
+    #     class ImageCategory(str, Enum):
+    #         """Enumeration for retail image categories."""
+    #         INK_WALL = "Ink Wall"
+    #         SHELVES_WITH_PRODUCTS = "Shelves with Products"
+    #         PRODUCTS_ON_FLOOR = "Products on Floor"
+    #         MERCHANDISING_ENDCAP = "Merchandising Endcap"
+    #         OTHER = "Other"
 
-        class ImageClassification(BaseModel):
-            """Schema for classifying a retail image."""
-            category: ImageCategory = Field(..., description="The best-fitting category for the image based on the provided definitions.")
-            confidence_score: float = Field(..., ge=0.0, le=1.0, description="The model's confidence in its classification, from 0.0 to 1.0.")
-            reasoning: str = Field(..., description="A brief explanation for why the image was assigned to this category.")
+    #     class ImageClassification(BaseModel):
+    #         """Schema for classifying a retail image."""
+    #         category: ImageCategory = Field(..., description="The best-fitting category for the image based on the provided definitions.")
+    #         confidence_score: float = Field(..., ge=0.0, le=1.0, description="The model's confidence in its classification, from 0.0 to 1.0.")
+    #         reasoning: str = Field(..., description="A brief explanation for why the image was assigned to this category.")
 
         classification_prompt = """
         You are an expert in retail image analysis. Your task is to classify the provided image into one of the following categories.
@@ -330,63 +334,63 @@ async def main():
 
         Analyze the image and provide your classification in the requested JSON format.
         """
-        image_path = BASE_DIR.joinpath('static', "be51ca05-802e-4dfd-bc53-fec65616d569-recap.jpeg")
-        classification_result = await client.ask_to_image(
-            image=image_path,
-            prompt=classification_prompt,
-            structured_output=ImageClassification,
-            model=GoogleModel.GEMINI_2_5_FLASH
-        )
-        print(classification_result)
-        if classification_result and isinstance(classification_result.output, ImageClassification):
-            # The code is assigning the output of a classification result to the variable `result` in
-            # Python.
-            result = classification_result.output
-            print("\n✅ Classification Complete:\n")
-            print(f"  - Category: {result.category.value}")
-            print(f"  - Confidence: {result.confidence_score:.2f}")
-            print(f"  - Reasoning: {result.reasoning}")
-            print("\n" + "-"*40)
+    #     image_path = BASE_DIR.joinpath('static', "be51ca05-802e-4dfd-bc53-fec65616d569-recap.jpeg")
+    #     classification_result = await client.ask_to_image(
+    #         image=image_path,
+    #         prompt=classification_prompt,
+    #         structured_output=ImageClassification,
+    #         model=GoogleModel.GEMINI_2_5_FLASH
+    #     )
+    #     print(classification_result)
+    #     if classification_result and isinstance(classification_result.output, ImageClassification):
+    #         # The code is assigning the output of a classification result to the variable `result` in
+    #         # Python.
+    #         result = classification_result.output
+    #         print("\n✅ Classification Complete:\n")
+    #         print(f"  - Category: {result.category.value}")
+    #         print(f"  - Confidence: {result.confidence_score:.2f}")
+    #         print(f"  - Reasoning: {result.reasoning}")
+    #         print("\n" + "-"*40)
 
-    async with GoogleGenAIClient() as client:
-        image_path = BASE_DIR.joinpath('static', "1bc3e9e8-3072-4c3a-8620-07fff9413a69-recap.jpeg")
-        shaq = BASE_DIR.joinpath('static', "Shaq.jpg")
-        response_shaq = await client.ask_to_image(
-            image=image_path,
-            reference_images=[shaq],
-            prompt=(
-                "First, analyze the image to see if Shaquille O'Neal (from the reference image) is present. "
-                "If he is, describe his location in the image, including any notable features or context. "
-                "Then, provide bounding boxes for his location. "
-                "Format your entire response as a single JSON object with 'analysis' and 'detections' keys."
-            ),
-            model=GoogleModel.GEMINI_2_5_FLASH,
-            count_objects=True
-        )
-        print("✅ analysis successful!")
-        result = response_shaq.output
-        print(f"Text Analysis: {result.analysis}")
-        print(f"Detections Found: {len(result.detections)}")
-        for box in result.detections:
-            print(f"  - Label: {box.label}, Box: {box.box_2d}")
-        # Product Detection
-        image_path = BASE_DIR.joinpath('static', "e1ad2662-c624-4f9b-a7aa-7518613a893a-recap.jpeg")
-        response_image = await client.ask_to_image(
-            image=image_path,
-            prompt=(
-                "Analyze the image methodically: identify and count the products in the image, even if it's partially visible."
-                " our main goal is to identify Epson products and competitive products in the image, "
-                " Epson competitors are: HP, Canon, Brother, Lexmark, Xerox, Dell, Samsung, Ricoh, Konica Minolta, Kyocera, Sharp, Toshiba, Panasonic."
-                " if there is no Epson or competitive products in the image, just say 'No Epson or competitive products found'. "
-                " if there are Epson or competitive products in the image, please provide the following information: "
-                "* Count number of products in the image (in box and unboxed) "
-                "* Provide a list of the products with their brand, model and type (printers, scanners, etc.)"
-                "* Provide a briefly description of the products "
-            ),
-            model=GoogleModel.GEMINI_2_5_FLASH,
-        )
-        print(response_image)
-        print("Image response text:", response_image.text)
+    # async with GoogleGenAIClient() as client:
+    #     image_path = BASE_DIR.joinpath('static', "1bc3e9e8-3072-4c3a-8620-07fff9413a69-recap.jpeg")
+    #     shaq = BASE_DIR.joinpath('static', "Shaq.jpg")
+    #     response_shaq = await client.ask_to_image(
+    #         image=image_path,
+    #         reference_images=[shaq],
+    #         prompt=(
+    #             "First, analyze the image to see if Shaquille O'Neal (from the reference image) is present. "
+    #             "If he is, describe his location in the image, including any notable features or context. "
+    #             "Then, provide bounding boxes for his location. "
+    #             "Format your entire response as a single JSON object with 'analysis' and 'detections' keys."
+    #         ),
+    #         model=GoogleModel.GEMINI_2_5_FLASH,
+    #         count_objects=True
+    #     )
+    #     print("✅ analysis successful!")
+    #     result = response_shaq.output
+    #     print(f"Text Analysis: {result.analysis}")
+    #     print(f"Detections Found: {len(result.detections)}")
+    #     for box in result.detections:
+    #         print(f"  - Label: {box.label}, Box: {box.box_2d}")
+    #     # Product Detection
+    #     image_path = BASE_DIR.joinpath('static', "e1ad2662-c624-4f9b-a7aa-7518613a893a-recap.jpeg")
+    #     response_image = await client.ask_to_image(
+    #         image=image_path,
+    #         prompt=(
+    #             "Analyze the image methodically: identify and count the products in the image, even if it's partially visible."
+    #             " our main goal is to identify Epson products and competitive products in the image, "
+    #             " Epson competitors are: HP, Canon, Brother, Lexmark, Xerox, Dell, Samsung, Ricoh, Konica Minolta, Kyocera, Sharp, Toshiba, Panasonic."
+    #             " if there is no Epson or competitive products in the image, just say 'No Epson or competitive products found'. "
+    #             " if there are Epson or competitive products in the image, please provide the following information: "
+    #             "* Count number of products in the image (in box and unboxed) "
+    #             "* Provide a list of the products with their brand, model and type (printers, scanners, etc.)"
+    #             "* Provide a briefly description of the products "
+    #         ),
+    #         model=GoogleModel.GEMINI_2_5_FLASH,
+    #     )
+    #     print(response_image)
+    #     print("Image response text:", response_image.text)
 
 
         # image_path = BASE_DIR.joinpath('static', "8455b202-28cf-4231-a9d9-7c175def0a93-recap.jpeg")
@@ -406,39 +410,39 @@ async def main():
         # print("Image response text:", response_image.text)
 
         # Ink-Wall Analysis:
-        class CleanlinessLevel(str, Enum):
-            """Enumeration for cleanliness condition."""
-            CLEAN = "Clean"
-            SLIGHTLY_DUSTY = "Slightly Dusty"
-            NEEDS_CLEANING = "Needs Cleaning"
+        # class CleanlinessLevel(str, Enum):
+        #     """Enumeration for cleanliness condition."""
+        #     CLEAN = "Clean"
+        #     SLIGHTLY_DUSTY = "Slightly Dusty"
+        #     NEEDS_CLEANING = "Needs Cleaning"
 
-        class InkWallAnalysis(BaseModel):
-            """Schema for a detailed retail ink-wall analysis."""
-            epson_product_count: int = Field(..., description="The total count of all visible Epson branded products (boxed or unboxed).")
-            competitor_product_count: int = Field(..., description="The total count of all visible competitor products (HP, Canon, Brother, etc.).")
-            competitors_found: List[str] = Field(..., description="A list of competitor brand names found on the shelves.")
-            shelf_occupancy_percentage: float = Field(..., description="An estimated percentage of how full the shelves are overall (0-100).")
-            overall_condition: str = Field(..., description="A brief description of the overall condition of the ink wall (e.g., 'well-stocked', 'needs restocking'), evaluate product alignment, cleanliness, and any visible damage")
-            has_gaps: bool = Field(..., description="True if there are noticeable empty spaces or gaps on the shelves where products should be.")
-            gaps_description: Optional[str] = Field(None, description="A brief description of where the most significant gaps are located.")
-            restocking_needed: bool = Field(..., description="True if the gaps or low stock levels suggest that restocking is required.")
-            is_shelf_empty: bool = Field(..., description="True if any single shelf is completely empty, otherwise False.")
-            misplaced_products_found: bool = Field(..., description="True if any product appears to be in the wrong location (e.g., an HP cartridge in an Epson section).")
-            cleanliness_condition: CleanlinessLevel = Field(..., description="The overall cleanliness of the shelves and products.")
-            trash_present: bool = Field(..., description="True if any visible trash, debris, or stray packaging is present on the shelves or floor.")
+        # class InkWallAnalysis(BaseModel):
+        #     """Schema for a detailed retail ink-wall analysis."""
+        #     epson_product_count: int = Field(..., description="The total count of all visible Epson branded products (boxed or unboxed).")
+        #     competitor_product_count: int = Field(..., description="The total count of all visible competitor products (HP, Canon, Brother, etc.).")
+        #     competitors_found: List[str] = Field(..., description="A list of competitor brand names found on the shelves.")
+        #     shelf_occupancy_percentage: float = Field(..., description="An estimated percentage of how full the shelves are overall (0-100).")
+        #     overall_condition: str = Field(..., description="A brief description of the overall condition of the ink wall (e.g., 'well-stocked', 'needs restocking'), evaluate product alignment, cleanliness, and any visible damage")
+        #     has_gaps: bool = Field(..., description="True if there are noticeable empty spaces or gaps on the shelves where products should be.")
+        #     gaps_description: Optional[str] = Field(None, description="A brief description of where the most significant gaps are located.")
+        #     restocking_needed: bool = Field(..., description="True if the gaps or low stock levels suggest that restocking is required.")
+        #     is_shelf_empty: bool = Field(..., description="True if any single shelf is completely empty, otherwise False.")
+        #     misplaced_products_found: bool = Field(..., description="True if any product appears to be in the wrong location (e.g., an HP cartridge in an Epson section).")
+        #     cleanliness_condition: CleanlinessLevel = Field(..., description="The overall cleanliness of the shelves and products.")
+        #     trash_present: bool = Field(..., description="True if any visible trash, debris, or stray packaging is present on the shelves or floor.")
 
 
-        image_path = BASE_DIR.joinpath('static', "a95ac304-8636-4772-99e1-4ed6fabaded1-recap.jpeg")
-        analysis_result = await client.ask_to_image(
-            image_path=image_path,
-            prompt="""
-            Analyze the provided image of a retail ink wall. Perform a detailed planogram and shelf-space analysis.
-            Use the following list of known competitors to identify non-Epson products: ['HP', 'Canon', 'Brother', 'Lexmark'].
-            Based on your visual analysis, populate all fields of the requested JSON schema.
-            """,
-            structured_output=InkWallAnalysis,
-            model=GoogleModel.GEMINI_2_5_FLASH
-        )
+        # image_path = BASE_DIR.joinpath('static', "a95ac304-8636-4772-99e1-4ed6fabaded1-recap.jpeg")
+        # analysis_result = await client.ask_to_image(
+        #     image_path=image_path,
+        #     prompt="""
+        #     Analyze the provided image of a retail ink wall. Perform a detailed planogram and shelf-space analysis.
+        #     Use the following list of known competitors to identify non-Epson products: ['HP', 'Canon', 'Brother', 'Lexmark'].
+        #     Based on your visual analysis, populate all fields of the requested JSON schema.
+        #     """,
+        #     structured_output=InkWallAnalysis,
+        #     model=GoogleModel.GEMINI_2_5_FLASH
+        # )
         # print(analysis_result)
         # if analysis_result and isinstance(analysis_result.output, InkWallAnalysis):
         #     result = analysis_result.output
