+++ parrot/models/responses.py
@@ -1,110 +1,28 @@
 from pathlib import Path
-from typing import Dict, List, Optional, Any, TypedDict, Callable
-from enum import Enum
+from typing import Dict, List, Optional, Any, TypedDict
 from datetime import datetime
-from dataclasses import dataclass
-from typing_extensions import Literal
 from pydantic import BaseModel, Field
+from .basic import CompletionUsage, ToolCall
 
 
-class OutputFormat(Enum):
-    """Supported output formats for structured responses."""
-    JSON = "json"
-    CSV = "csv"
-    YAML = "yaml"
-    XML = "xml"
-    CODE = "code"
-    CUSTOM = "custom"
-    TEXT = "text"
+class StreamChunk(BaseModel):
+    """Represents a chunk in a streaming response."""
+    content: str
+    is_complete: bool = False
+    chunk_id: Optional[str] = None
+    turn_id: Optional[str] = None
 
 
-@dataclass
-class StructuredOutputConfig:
-    """Configuration for structured output parsing."""
-    output_type: type
-    format: OutputFormat = OutputFormat.JSON
-    custom_parser: Optional[Callable[[str], Any]] = None
-
-
-
-class ToolCall(BaseModel):
-    """Unified tool call representation."""
+class MessageResponse(TypedDict):
+    """Response structure for LLM messages."""
     id: str
-    name: str
-    arguments: Dict[str, Any]
-    result: Optional[Any] = None
-    error: Optional[str] = None
-    execution_time: Optional[float] = None
-
-
-class CompletionUsage(BaseModel):
-    """Unified completion usage tracking across different LLM providers."""
-
-    # Core usage metrics (common across all providers)
-    prompt_tokens: int = 0
-    completion_tokens: int = 0
-    total_tokens: int = 0
-
-    # Timing information (optional, provider-specific)
-    completion_time: Optional[float] = None
-    prompt_time: Optional[float] = None
-    queue_time: Optional[float] = None
-    total_time: Optional[float] = None
-
-    # Cost information (optional)
-    estimated_cost: Optional[float] = None
-
-    # Provider-specific additional fields
-    extra_usage: Dict[str, Any] = Field(default_factory=dict)
-
-    @classmethod
-    def from_openai(cls, usage: Any) -> "CompletionUsage":
-        """Create from OpenAI usage object."""
-        return cls(
-            prompt_tokens=getattr(usage, 'prompt_tokens', 0),
-            completion_tokens=getattr(usage, 'completion_tokens', 0),
-            total_tokens=getattr(usage, 'total_tokens', 0)
-        )
-
-    @classmethod
-    def from_groq(cls, usage: Any) -> "CompletionUsage":
-        """Create from Groq usage object."""
-        return cls(
-            prompt_tokens=getattr(usage, 'prompt_tokens', 0),
-            completion_tokens=getattr(usage, 'completion_tokens', 0),
-            total_tokens=getattr(usage, 'total_tokens', 0),
-            completion_time=getattr(usage, 'completion_time', None),
-            prompt_time=getattr(usage, 'prompt_time', None),
-            queue_time=getattr(usage, 'queue_time', None),
-            total_time=getattr(usage, 'total_time', None)
-        )
-
-    @classmethod
-    def from_claude(cls, usage: Dict[str, Any]) -> "CompletionUsage":
-        """Create from Claude usage dict."""
-        return cls(
-            prompt_tokens=usage.get('input_tokens', 0),
-            completion_tokens=usage.get('output_tokens', 0),
-            total_tokens=usage.get('input_tokens', 0) + usage.get('output_tokens', 0),
-            extra_usage=usage
-        )
-
-    @classmethod
-    def from_gemini(cls, usage: Dict[str, Any]) -> "CompletionUsage":
-        """Create from Gemini/Vertex AI usage dict."""
-        # Handle both Gemini API format and Vertex AI format
-        prompt_tokens = usage.get('prompt_token_count', 0) or usage.get('prompt_tokens', 0)
-        completion_tokens = usage.get(
-            'candidates_token_count', 0
-        ) or usage.get('completion_tokens', 0)
-        total_tokens = usage.get('total_token_count', 0) or usage.get('total_tokens', 0)
-
-        return cls(
-            prompt_tokens=prompt_tokens,
-            completion_tokens=completion_tokens,
-            total_tokens=total_tokens,
-            extra_usage=usage
-        )
+    type: str
+    role: str
+    content: List[Dict[str, Any]]
+    model: str
+    stop_reason: Optional[str]
+    stop_sequence: Optional[str]
+    usage: Dict[str, int]
 
 
 class AIMessage(BaseModel):
@@ -192,7 +110,7 @@ class AIMessage(BaseModel):
         arbitrary_types_allowed = True
 
     @property
-    def text(self) -> str:
+    def to_text(self) -> str:
         """Get text representation of output."""
         if isinstance(self.output, str):
             return self.output
@@ -206,6 +124,7 @@ class AIMessage(BaseModel):
             # Handle pandas DataFrames
             return self.output.to_string()
         else:
+            # Return the output *as is*
             return str(self.output)
 
     @property
@@ -222,25 +141,6 @@ class AIMessage(BaseModel):
         """Add a tool call to the response."""
         self.tool_calls.append(tool_call)  # pylint: disable=E1101 # noqa
 
-
-class StreamChunk(BaseModel):
-    """Represents a chunk in a streaming response."""
-    content: str
-    is_complete: bool = False
-    chunk_id: Optional[str] = None
-    turn_id: Optional[str] = None
-
-
-class MessageResponse(TypedDict):
-    """Response structure for LLM messages."""
-    id: str
-    type: str
-    role: str
-    content: List[Dict[str, Any]]
-    model: str
-    stop_reason: Optional[str]
-    stop_sequence: Optional[str]
-    usage: Dict[str, int]
 
 # Factory functions to create AIMessage from different providers
 class AIMessageFactory:
@@ -418,62 +318,3 @@ class AIMessageFactory:
     @staticmethod
     def from_video(**kwargs):
         return AIMessage(**kwargs)
-
-
-class BoundingBox(BaseModel):
-    """Represents a detected object with its location and details."""
-    object_id: str = Field(..., description="Unique identifier for this detection")
-    brand: str = Field(..., description="Product brand (Epson, HP, Canon, etc.)")
-    model: Optional[str] = Field(None, description="Product model if identifiable")
-    product_type: str = Field(..., description="Type of product (printer, scanner, ink cartridge, etc.)")
-    description: str = Field(..., description="Brief description of the product")
-    confidence: float = Field(..., description="Confidence level (0.0 to 1.0)")
-    # Simple bounding box as [x1, y1, x2, y2] normalized coordinates (0.0 to 1.0)
-    bbox: List[float] = Field(..., description="Bounding box coordinates [x1, y1, x2, y2] as normalized values (0.0-1.0)")
-
-
-class ObjectDetectionResult(BaseModel):
-    """A list of all prominent items detected in the image."""
-    analysis: str = Field(
-        ...,
-        description="A detailed text analysis of the image that answers the user's prompt."
-    )
-    total_count: int = Field(..., description="Total number of products detected")
-    detections: List[BoundingBox] = Field(
-        default_factory=list,
-        description="A list of bounding boxes for all prominent detected objects."
-    )
-
-class ImageGenerationPrompt(BaseModel):
-    """Input schema for generating an image."""
-    prompt: str = Field(..., description="The main text prompt describing the desired image.")
-    styles: Optional[List[str]] = Field(default_factory=list, description="Optional list of styles to apply (e.g., 'photorealistic', 'cinematic', 'anime').")
-    model: str = Field(description="The image generation model to use.")
-    negative_prompt: Optional[str] = Field(None, description="A description of what to avoid in the image.")
-    aspect_ratio: str = Field(default="1:1", description="The desired aspect ratio (e.g., '1:1', '16:9', '9:16').")
-
-
-class SpeakerConfig(BaseModel):
-    """Configuration for a single speaker in speech generation."""
-    name: str = Field(..., description="The name of the speaker in the script (e.g., 'Joe', 'Narrator').")
-    voice: str = Field(..., description="The pre-built voice name to use (e.g., 'Kore', 'Puck', 'Chitose').")
-    # Gender is often inferred from the voice, but can be included for clarity
-    gender: Optional[str] = Field(None, description="The gender associated with the voice (e.g., 'Male', 'Female').")
-
-
-class SpeechGenerationPrompt(BaseModel):
-    """Input schema for generating speech from text."""
-    prompt: str = Field(..., description="The text to be converted to speech. For multiple speakers, use their names (e.g., 'Joe: Hello. Jane: Hi there.').")
-    speakers: List[SpeakerConfig] = Field(..., description="A list of speaker configurations. Use one for a single voice, multiple for a conversation.")
-    model: Optional[str] = Field(default=None, description="The text-to-speech model to use.")
-    language: Optional[str] = Field("en-US", description="Language code for the conversation.")
-
-
-class VideoGenerationPrompt(BaseModel):
-    """Input schema for generating video content."""
-    prompt: str = Field(..., description="The text prompt describing the desired video content.")
-    number_of_videos: int = Field(default=1, description="The number of videos to generated per request.")
-    model: str = Field(..., description="The video generation model to use.")
-    aspect_ratio: str = Field(default="16:9", description="The desired aspect ratio (e.g., '16:9', '9:16').")
-    duration: Optional[int] = Field(None, description="Optional duration in seconds for the video.")
-    negative_prompt: Optional[str] = Field(default='', description="A description of what to avoid in the video.")
